example/requirements.txt
# Your app requirements.
-r ../requirements_test.txt

# Your app in editable mode.
-e ../

Django==1.11
psycopg2==2.7.1
django-redis-cache==1.7.1
hiredis==0.2.0
redis==2.10.5
django-celery-beat==1.5.0
django_busybody==0.1.30
